clean repeated codes in classifier
